# perf(statusTile): faster array creation and lodash lazy loading