--- a/.all-contributorsrc
+++ b/.all-contributorsrc
@@ -246,6 +246,16 @@
       "contributions": [
         "code"
       ]
+    },
+    {
+      "login": "haroldtreen",
+      "name": "Harold Treen",
+      "avatar_url": "https://avatars2.githubusercontent.com/u/1745854?v=4",
+      "profile": "https://haroldtreen.com",
+      "contributions": [
+        "bug",
+        "code"
+      ]
     }
   ]
 }