### Description

Reduces the activation time for `prettier-atom` from 248ms down to ~60ms.

![image](https://user-images.githubusercontent.com/1745854/34460337-3e48127a-edd9-11e7-9284-e140719cad30.png)

### Changes

- Removes the array spread operator from `atomInterface`. This stops `babel-runtime` from requiring large polyfills (`babel-runtime/helpers/toConsumableArray`). Saves ~70ms.
- Lazy loads `lodash/fp/flow` in `editorInterface`. This function is only used for `getCurrentDir`, but gets loaded whenever `editorInterface` is loaded anywhere. Save ~120ms.